Voxelization writes data to 3D texture, had to specify bind image with layered = true to write to more than first layer (not very clear in the docs)

+++ app/src/main/assets/shaders/voxelization.frag
@@ -121,7 +121,7 @@ void main()
 
 	ivec3 voxelCoord;
 	float res = float(scene.voxelRes);
-	int depthCoord = int(gl_FragCoord.z * res);
+	float depthCoord = gl_FragCoord.z * res;
 
 
 	if(domInd == uint(0)) {
